% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_scNetwork_degree.R
\name{calculate_scNetwork_degree}
\alias{calculate_scNetwork_degree}
\title{Title}
\usage{
calculate_scNetwork_degree(
  scNetworks,
  top,
  cell_type_annotation,
  ntype = NULL,
  column_name = "undup.sub.celltype"
)
}
\arguments{
\item{column_name}{}
}
\description{
Title
}
