% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_scNetwork_precision_recall.R
\name{Calculate_scNetwork_precision_recall}
\alias{Calculate_scNetwork_precision_recall}
\title{Calculate the precision and recall of single-cell specific network}
\usage{
Calculate_scNetwork_precision_recall(
  scNetworks,
  TF_target_pair,
  top_number = c(1000, seq(2000, 10000, by = 2000), 20000),
  gene_id_gene_name_pair = NULL,
  gene_name_type = NULL
)
}
\arguments{
\item{gene_name_type}{}
}
\description{
Calculate the precision and recall of single-cell specific network
}
