% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_scNetwork_precision_recall_top.R
\name{Calculate_scNetwork_precision_recall_top}
\alias{Calculate_scNetwork_precision_recall_top}
\title{Calculate the precision and recall of top gene pairs from single-cell specific network}
\usage{
Calculate_scNetwork_precision_recall_top(scNet_precision_recall, top_number)
}
\arguments{
\item{top_number}{}
}
\description{
Calculate the precision and recall of top gene pairs from single-cell specific network
}
