% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_wScReNI_scNetworks.R
\name{Combine_wScReNI_scNetworks}
\alias{Combine_wScReNI_scNetworks}
\title{combine single-cell regulatory networks from wScReNI}
\usage{
Combine_wScReNI_scNetworks(sub.scatac.top, network.path, cell.index = NULL)
}
\arguments{
\item{cell.index}{}
}
\description{
combine single-cell regulatory networks from wScReNI
}
