% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_module.R
\name{enrich_module}
\alias{enrich_module}
\title{Title}
\usage{
enrich_module(
  Kmeans_result,
  org.db,
  enrich.db,
  fun_num = 5,
  pvalueCutoff = 0.05,
  use_internal_data = TRUE,
  organism = NULL
)
}
\arguments{
\item{organism}{}
}
\description{
Title
}
