% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CACIMAR_R_NewFunctions.R
\name{filter_regulation_fimo}
\alias{filter_regulation_fimo}
\title{Title}
\usage{
filter_regulation_fimo(fimo_regulation, regulatory_relationships)
}
\arguments{
\item{regulatory_relationships}{}
}
\description{
Title
}
