% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wScReNI_affiliated_functions.R
\name{gene_peak_corr1}
\alias{gene_peak_corr1}
\title{Title}
\usage{
gene_peak_corr1(
  exprMatrix,
  scatac,
  peak_gene_overlap_GR,
  threshold = 0.1,
  nthread = 20
)
}
\arguments{
\item{downstream_len}{}
}
\description{
Title
}
