% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wScReNI_affiliated_functions.R
\name{Get_motif_peak_pair_df0}
\alias{Get_motif_peak_pair_df0}
\title{Title}
\usage{
Get_motif_peak_pair_df0(
  peak_gene_overlap_GR2,
  gene_name_type,
  motif_database,
  motif_pwm,
  genome_database,
  pvalue = 5^(-4),
  nthread = 10
)
}
\description{
Title
}
