% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_scRNA_scATAC_neighbors.R
\name{Get_scRNA_scATAC_neighbors}
\alias{Get_scRNA_scATAC_neighbors}
\title{Title}
\usage{
Get_scRNA_scATAC_neighbors(
  coembed,
  celltypes = "celltypes",
  datatypes = "datatypes"
)
}
\arguments{
\item{datatypes}{}
}
\description{
Title
}
