% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_enriched_scRegulators.R
\name{Identify_enriched_scRegulators}
\alias{Identify_enriched_scRegulators}
\title{Identify regulators enriched in each single-cell regulatory network}
\usage{
Identify_enriched_scRegulators(
  regulatory_network,
  Kmeans_clustering_ENS,
  TFFDR1 = 10,
  TFFDR2 = 10
)
}
\arguments{
\item{TFFDR2}{}
}
\description{
Identify regulators enriched in each single-cell regulatory network
}
