% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_CSN_scNetworks.R
\name{Infer_CSN_scNetworks}
\alias{Infer_CSN_scNetworks}
\title{Title}
\usage{
Infer_CSN_scNetworks(data, alpha = 0.01, boxsize = 0.1, weighted = 0)
}
\arguments{
\item{weighted}{}
}
\description{
Title
}
