% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_gene_peak_relationships.R
\name{Infer_gene_peak_relationships}
\alias{Infer_gene_peak_relationships}
\title{Title}
\usage{
Infer_gene_peak_relationships(
  gtf_data,
  scrna,
  scatac,
  motif_database,
  motif_pwm,
  genome_database,
  gene_name_type = "symbol",
  upstream_len = 1000 * 250,
  downstream_len = 1000 * 250,
  threshold = 0.1,
  nthread = 20,
  pvalue = 5^(-4)
)
}
\arguments{
\item{pvalue}{}
}
\description{
Title
}
