% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_kScReNI_scNetworks.R
\name{Infer_kScReNI_scNetworks}
\alias{Infer_kScReNI_scNetworks}
\title{Infer single-cell specific networks using kScReNI}
\usage{
Infer_kScReNI_scNetworks(
  exprMatrix,
  nfeatures = 4000,
  knn = 20,
  nthread = 20,
  nTrees = 100
)
}
\arguments{
\item{nTrees}{}
}
\description{
Infer single-cell specific networks using kScReNI
}
