% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_LIONESS_scNetworks.R
\name{Infer_LIONESS_scNetworks}
\alias{Infer_LIONESS_scNetworks}
\title{Infer single-cell specific networks using LIONESS}
\usage{
Infer_LIONESS_scNetworks(exprMatrix, nCores = 60, nTrees = 100)
}
\arguments{
\item{nTrees}{}
}
\description{
Infer single-cell specific networks using LIONESS
}
