% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_wScReNI_scNetworks.R
\name{Infer_wScReNI_scNetworks}
\alias{Infer_wScReNI_scNetworks}
\title{Construct single-cell regulatory networks using wScReNI}
\usage{
Infer_wScReNI_scNetworks(
  exprMatrix,
  gene_peak_overlap_matrix,
  gene_peak_overlap_labs,
  nearest.neighbors.idx,
  network.path,
  data.name,
  cell.index = NULL,
  nthread = 50,
  max.cell.per.batch = 10
)
}
\arguments{
\item{max.cell.per.batch}{}
}
\description{
Construct single-cell regulatory networks using wScReNI
}
