% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate_scRNA_scATAC.R
\name{Integrate_scRNA_scATAC}
\alias{Integrate_scRNA_scATAC}
\title{Title}
\usage{
Integrate_scRNA_scATAC(
  scRNAseq,
  scATACseq,
  AnchorsDim = 50,
  IntegratedDimensions,
  KNN = 20,
  data.type = "unpaired",
  species = "mouse"
)
}
\arguments{
\item{species}{}
}
\description{
Title
}
