% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wScReNI_affiliated_functions.R
\name{peak_gene_overlap_GR1}
\alias{peak_gene_overlap_GR1}
\title{Title}
\usage{
peak_gene_overlap_GR1(
  gtf_data,
  scrna,
  gene_name_type,
  scatac,
  upstream_len,
  downstream_len
)
}
\arguments{
\item{scatac}{}

\item{downstream_len}{}
}
\description{
Title
}
