% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wScReNI_affiliated_functions.R
\name{peak_gene_TF_match}
\alias{peak_gene_TF_match}
\title{Title}
\usage{
peak_gene_TF_match(motif_database, motif_peak_pair_df, peak_gene_overlap_GR2)
}
\arguments{
\item{peak_gene_overlap_GR2}{}
}
\description{
Title
}
