% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_integrated_scRNA_scATAC.R
\name{Plot_integrated_scRNA_scATAC}
\alias{Plot_integrated_scRNA_scATAC}
\title{Title}
\usage{
Plot_integrated_scRNA_scATAC(
  coembed,
  figure.folder = result.path,
  data.name,
  data.type = "unpaired",
  groups = c("datatypes", "samples", "celltypes"),
  genes = "Rlbp1"
)
}
\arguments{
\item{genes}{}
}
\description{
Title
}
