% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_scNetwork_clustering.R
\name{Plot_scNetwork_clustering}
\alias{Plot_scNetwork_clustering}
\title{Perform clustering based on the gene degrees of cell-specific networks}
\usage{
Plot_scNetwork_clustering(
  degree_all,
  path,
  data.name,
  cell_type_annotation,
  col,
  cell_type_column
)
}
\arguments{
\item{cell_type_column}{}
}
\description{
Perform clustering based on the gene degrees of cell-specific networks
}
