% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{select_features}
\alias{select_features}
\title{Title}
\usage{
select_features(scrna, nfeatures = 4000, datatype = "RNA", SeuratLayer = T)
}
\arguments{
\item{SeuratLayer}{}
}
\description{
Title
}
