% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_partial_cells_for_scNewtorks.R
\name{Select_partial_cells_for_scNewtorks}
\alias{Select_partial_cells_for_scNewtorks}
\title{Title}
\usage{
Select_partial_cells_for_scNewtorks(
  coembed,
  Partial_cell,
  cell.num,
  data.type = "unpaired",
  scRNA_scATAC_neighbor_undup = NULL,
  cell.name = NULL,
  Celltypename = "undup.cell.types"
)
}
\arguments{
\item{Celltypename}{}
}
\description{
Title
}
